requirements.txt
django==4.2.3
djangorestframework==3.14.0
djangorestframework-authtoken
gunicorn==21.2.0